+++ packages/database/src/database/header.js
@@ -1,14 +1,5 @@
-const name = 'home-gallery/database'
-const major = 1
-const minor = 3
+import { GalleryFileType } from '@home-gallery/common'
 
-export const isDatabaseTypeCompatible = type => type && type.startsWith(`${name}@${major}.`)
+export const databaseFileType = new GalleryFileType('home-gallery/database@1.3')
 
-export const HeaderType = `${name}@${major}.${minor}`
-
-export const getMajorMinor = type => {
-  if (!type.startsWith(`${name}@`)) {
-    throw new Error(`Invalid type. Expecting ${name}@... but was ${type}`)
-  }
-  return type.substring(name.length + 1).split('.').map(v => +v)
-}
+export const HeaderType = `${databaseFileType}`