+++ packages/database/src/database/index.js
@@ -1,4 +1,4 @@
-export { HeaderType, isDatabaseTypeCompatible } from './header.js'
+export * from './header.js'
 export { initDatabase, readDatabase, readOrCreateDatabase } from './read-database.js'
 export { writeDatabase, writeDatabasePlain } from './write-database.js'
 export { migrate } from './migrate.js'